feat: add custom access middleware for Filament
