+++ app/Http/Middleware/FilamentAccessMiddleware.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+
+class FilamentAccessMiddleware
+{
+    /**
+     * Handle an incoming request.
+     *
+     * @param  \Closure(\Illuminate\Http\Request): (\Illuminate\Http\Response|\Illuminate\Http\RedirectResponse)  $next
+     * @return \Illuminate\Http\Response|\Illuminate\Http\RedirectResponse
+     */
+    public function handle(Request $request, Closure $next)
+    {
+        $account = $request->user();
+
+        if (! $account->can('use-permission', 'admin/beta')) {
+            return abort(403);
+        }
+
+        return $next($request);
+    }
+}
